#34-fix-dto

+++ src/main/java/com/example/Dynamo_Backend/dto/RequestDto/MachineRequestDto.java
@@ -18,7 +18,7 @@ public class MachineRequestDto {
     private Integer month;
     private Float oee;
     private String groupId;
-    private Integer machineMiningTarget;
+    private Float machineMiningTarget;
     private String createdDate;
     private String updatedDate;
 }
